Added BasicPhysicsObject, SpringForces
Removed TestPhysicsObject, and replaced it with BasicPhysicsObject
(which will exist in the v 1.0 release).

Also added BasicSpringForce, AnchoredSpringForce, and PlayerMotionForce,
three of the forces intended to come with v. 1.0

WorldManager now has a constructor that will construct the
ICollisionManager object to one of the bundled collision manager types.
Right now, only BVHTree is supported.

Updated Vect3Normal documentation.